requirements.txt
discord.py
aiohttp
beautifulsoup4
pytz
motor
pymongo
python-dotenv